Crate demo2

Ratatui Demo2 example

The latest version of this example is available in the examples folder in the repository.

Please note that the examples are designed to be run against the main branch of the Github repository. This means that you may not be able to compile with the latest release version on crates.io, or the one that you have installed locally.

See the examples readme for more information on finding examples that match the version of the library you are using.

Modules

app 🔒

colors 🔒

destroy 🔒

tabs 🔒

theme 🔒

Structs

RgbSwatch

A widget that renders a color swatch of RGB colors.

Constants

THEME

Functions

color_from_oklab

Convert a hue and value into an RGB color via the Oklab color space.

main 🔒